% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namsor_endpoints.R
\name{namsor_country}
\alias{namsor_country}
\title{Obtain country of residence estimation for a name}
\usage{
namsor_country(firstName = "", lastName)
}
\arguments{
\item{firstName}{First name}

\item{lastName}{Last name}
}
\value{
Returns the most likely geographic context of a full name (first name and last name) or a surname.
}
\description{
Obtain country of residence estimation for a name
}
\examples{
\dontrun{namsor_country('John','Doe')}
}
