% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{namsor_country_to_iso}
\alias{namsor_country_to_iso}
\title{Country names to ISO codes conversion list}
\format{
named list with country names as slot names and ISO codes as values
}
\usage{
namsor_country_to_iso
}
\description{
Country names to ISO codes conversion list
}
\keyword{datasets}
