% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namsor_endpoints.R
\name{namsor_diaspora}
\alias{namsor_diaspora}
\title{Obtain country of residence estimation for a name}
\usage{
namsor_diaspora(firstName, lastName, cntryResid)
}
\arguments{
\item{firstName}{First name}

\item{lastName}{Last name}

\item{cntryResid}{Country of residence as a 2-character ISO code. See namsor::namsor_country_to_iso for options}
}
\value{
Returns the most likely ethnicity or diaspora of a first name and last name, according to the country of residence.
}
\description{
Obtain country of residence estimation for a name
}
\examples{
\dontrun{namsor_diaspora('John','Doe', 'BE')}
}
