% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namsor_endpoints.R
\name{namsor_gender}
\alias{namsor_gender}
\title{Obtain gender estimation for a name}
\usage{
namsor_gender(firstName, lastName)
}
\arguments{
\item{firstName}{First name}

\item{lastName}{Last name}
}
\value{
Returns the most likely gender of a first name and a last name.
}
\description{
Obtain gender estimation for a name
}
\examples{
\dontrun{namsor_gender('John','Doe')}
}
