% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{namsor_iso_to_country}
\alias{namsor_iso_to_country}
\title{ISO to country names  codes conversion list}
\format{
named list with ISO codes as slot names and country names as values
}
\usage{
namsor_iso_to_country
}
\description{
ISO to country names  codes conversion list
}
\keyword{datasets}
