% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namsor_endpoints.R
\name{namsor_origin}
\alias{namsor_origin}
\title{Obtain country of origin estimation for a name}
\usage{
namsor_origin(firstName, lastName)
}
\arguments{
\item{firstName}{First name}

\item{lastName}{Last name}
}
\value{
Returns the most likely country of origin of a first name and a last name. Note that the “Diaspora” endpoint may be better suited for multicultural countries like the U.S.A., Canada, Russia, etc.
}
\description{
Obtain country of origin estimation for a name
}
\examples{
\dontrun{namsor_origin('John','Doe')}
}
